refactor: extract TeamGraph and VisualSide into _components
