+++ src/components/LoginForm/LoginForm.tsx
@@ -20,277 +20,15 @@ import {
 } from '@mui/material'
 import Link from 'next/link'
 import { supabase } from '@/lib/supabase'
-import GrimChip from '@/components/GrimChip'
 import ConstellationMark from '@/components/TopNav/_components/ConstellationMark'
 import { EyeIcon, EyeOffIcon, GoogleGlyph, GitHubGlyph } from './_components/Icons'
+import VisualSide from './_components/VisualSide'
 
 const schema = z.object({
   email: z.string().min(1, 'Email is required').email({ message: 'Enter a valid email' }),
   password: z.string().min(1, 'Password is required'),
 })
 type FormValues = z.infer<typeof schema>
-
-function TeamGraph() {
-  const { palette } = useTheme()
-  const W = 560,
-    H = 480
-  const cx = W / 2,
-    cy = H / 2
-  type NodeStatus = 'complete' | 'running' | 'waiting' | 'idle'
-  const nodes: {
-    id: string
-    label: string
-    role: string
-    x: number
-    y: number
-    r: number
-    status: NodeStatus
-  }[] = [
-    {
-      id: 'coord',
-      label: 'Coordinator',
-      role: 'orchestrates',
-      x: cx,
-      y: cy,
-      r: 60,
-      status: 'running',
-    },
-    {
-      id: 'fetch',
-      label: 'Fetcher',
-      role: 'finds listings',
-      x: cx - 180,
-      y: cy - 110,
-      r: 46,
-      status: 'complete',
-    },
-    {
-      id: 'fit',
-      label: 'Analyser',
-      role: 'scores fit',
-      x: cx + 180,
-      y: cy - 110,
-      r: 46,
-      status: 'complete',
-    },
-    {
-      id: 'tail',
-      label: 'Tailor',
-      role: 'rewrites resume',
-      x: cx,
-      y: cy + 170,
-      r: 46,
-      status: 'running',
-    },
-  ]
-  const NODE_COLORS: Record<'complete' | 'running' | 'waiting' | 'idle', string> = {
-    complete: palette.secondary.main,
-    running: palette.primary.light,
-    waiting: palette.warning.main,
-    idle: palette.text.secondary,
-  }
-  const nodeColor = (status: 'complete' | 'running' | 'waiting' | 'idle'): string =>
-    NODE_COLORS[status]
-
-  return (
-    <Box
-      sx={{
-        bgcolor: 'background.paper',
-        border: '1px solid',
-        borderColor: 'divider',
-        borderRadius: 3,
-        p: 2,
-        boxShadow: '0 30px 80px -20px rgba(0,0,0,0.55)',
-        width: '100%',
-      }}
-    >
-      <Box
-        sx={{
-          display: 'flex',
-          alignItems: 'center',
-          justifyContent: 'space-between',
-          px: 1.25,
-          pb: 1.75,
-        }}
-      >
-        <Typography
-          variant="caption"
-          sx={{ color: 'text.secondary', fontFamily: 'ui-monospace, monospace' }}
-        >
-          team · job-search
-        </Typography>
-        <GrimChip color="secondary">
-          <Box
-            component="span"
-            sx={{
-              width: 6,
-              height: 6,
-              borderRadius: '50%',
-              bgcolor: 'secondary.main',
-              flexShrink: 0,
-            }}
-          />
-          Running
-        </GrimChip>
-      </Box>
-
-      <svg viewBox={`0 0 ${W} ${H}`} width="100%" style={{ display: 'block' }}>
-        <defs>
-          <radialGradient id="loginCoordGlow" cx="50%" cy="50%" r="50%">
-            <stop offset="0%" stopColor={palette.secondary.main} stopOpacity="0.35" />
-            <stop offset="100%" stopColor={palette.secondary.main} stopOpacity="0" />
-          </radialGradient>
-        </defs>
-
-        {/* edges */}
-        {nodes.slice(1).map((n) => (
-          <line
-            key={n.id}
-            x1={cx}
-            y1={cy}
-            x2={n.x}
-            y2={n.y}
-            stroke={palette.divider}
-            strokeWidth="1.5"
-            strokeDasharray="4 4"
-            style={{ animation: 'grimOrbit 3s linear infinite' }}
-          />
-        ))}
-
-        {/* centre glow */}
-        <circle cx={cx} cy={cy} r="120" fill="url(#loginCoordGlow)" />
-
-        {/* nodes */}
-        {nodes.map((n) => {
-          const c = nodeColor(n.status)
-          return (
-            <g
-              key={n.id}
-              style={{
-                animation:
-                  n.status === 'running' ? 'grimNodePulse 2.2s ease-in-out infinite' : 'none',
-              }}
-            >
-              <circle
-                cx={n.x}
-                cy={n.y}
-                r={n.r}
-                fill={palette.background.default}
-                stroke={c}
-                strokeWidth="1.5"
-              />
-              <circle
-                cx={n.x}
-                cy={n.y}
-                r={n.r + 6}
-                fill="none"
-                stroke={c}
-                strokeWidth="1"
-                opacity="0.25"
-              />
-              <text
-                x={n.x}
-                y={n.y - 5}
-                textAnchor="middle"
-                fontFamily="Inter, sans-serif"
-                fontSize="15"
-                fontWeight="600"
-                fill={palette.text.primary}
-              >
-                {n.label}
-              </text>
-              <text
-                x={n.x}
-                y={n.y + 14}
-                textAnchor="middle"
-                fontFamily="Inter, sans-serif"
-                fontSize="12"
-                fill={palette.text.secondary}
-              >
-                {n.role}
-              </text>
-            </g>
-          )
-        })}
-
-        {/* travelling packets */}
-        {nodes.slice(1).map((n, i) => (
-          <circle key={'p' + n.id} r="3" fill={nodeColor(n.status)}>
-            <animateMotion
-              dur={`${2 + i * 0.4}s`}
-              repeatCount="indefinite"
-              path={`M${cx},${cy} L${n.x},${n.y}`}
-            />
-            <animate
-              attributeName="opacity"
-              values="0;1;0"
-              dur={`${2 + i * 0.4}s`}
-              repeatCount="indefinite"
-            />
-          </circle>
-        ))}
-      </svg>
-    </Box>
-  )
-}
-
-function VisualSide() {
-  const { palette } = useTheme()
-  return (
-    /* Outer: stretches to grid row height, never grows beyond it */
-    <Box
-      sx={{
-        position: 'relative',
-        overflow: 'hidden',
-        borderLeft: '1px solid',
-        borderColor: 'divider',
-        alignSelf: 'stretch',
-      }}
-    >
-      {/* Inner: fills the outer, never adds to it */}
-      <Box
-        sx={{
-          position: 'absolute',
-          inset: 0,
-          overflow: 'hidden',
-          background: `linear-gradient(180deg, ${palette.background.default} 0%, #0A0816 100%)`,
-          display: 'flex',
-          flexDirection: 'column',
-          alignItems: 'center',
-          justifyContent: 'center',
-          p: '48px 56px',
-        }}
-      >
-        {/* Ambient mint glow */}
-        <Box
-          sx={{
-            position: 'absolute',
-            right: -150,
-            top: '35%',
-            width: 700,
-            height: 700,
-            borderRadius: '50%',
-            background: `radial-gradient(closest-side, ${palette.secondary.main}1f, transparent 70%)`,
-            pointerEvents: 'none',
-          }}
-        />
-        <Box sx={{ position: 'relative', zIndex: 1, width: '100%', maxWidth: 520 }}>
-          <Typography
-            variant="h2"
-            sx={{ fontSize: 36, lineHeight: 1.1, letterSpacing: '-0.02em', mb: 1 }}
-          >
-            Your agents are already working.
-          </Typography>
-          <Typography variant="body1" color="text.secondary" sx={{ mb: 3.5 }}>
-            A coordinator, a fetcher, an analyser, a tailor — running in parallel while you sleep.
-            Log back in and pick up where they left off.
-          </Typography>
-          <TeamGraph />
-        </Box>
-      </Box>
-    </Box>
-  )
-}
 
 export default function LoginForm() {
   const router = useRouter()
@@ -565,7 +303,6 @@ export default function LoginForm() {
         </Box>
       </Box>
 
-      {/* Visual side */}
       <VisualSide />
     </Box>
   )
