+++ src/components/LoginForm/_components/TeamGraph/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TeamGraph'
