+++ src/components/LoginForm/_components/VisualSide/index.ts
@@ -1,0 +1,1 @@
+export { default } from './VisualSide'
